editoast: isolate timetable CLI into its own module
